+++ lib/CHANGELOG.md
@@ -1,3 +1,7 @@
+# <unreleased>
+
+ * Improved efficiency of `is_valid` for option sets
+
 # 1.3.0 (October 29th, 2018)
 
 **Behaviour change:** This release contains an insignificant behaviour change with respect to